WIP add CellposeFlowDynamics op
